refactor and merge data
