--- src/data/shuzoData.json
@@ -1,44 +1,0 @@
-[
-  {
-    "id": "ANTELOPE",
-    "name": "ANTELOPE",
-    "title": "",
-    "description": "ANTELOPE酒廠，位於日本滋賀縣，由谷澤優気和福井駿於2020年創立，專注於製作Mead，一種蜂蜜發酵酒。谷澤優気在京都大学農学部學習期間開始接觸克拉夫特啤酒製作，並在静岡學習酒的醸造。他們的酒款多樣化，使用了如粉紅葡萄柚和辣椒等副原料，創造出不同風味的Mead。ANTELOPE注重與顧客和社區的連結，他們的理念是透過酒精飲品連接人與人之間的緣分，創造美妙的體驗。Mead，被認為是世界上最古老的酒精飲料之一，已存在了8000-9000年。由蜂蜜、水和酵母製成，蜜酒在古希臘、古埃及和北歐神話中佔有重要地位，曾是皇室和貴族的專屬飲品，並在古典文學中提及。傳統蜜酒（Traditional Mead）僅用蜂蜜、水和酵母製成，而手工蜜酒（Craft Mead）則加入各式水果、香料和植物性成分，展現多樣化口感。這種既古老又新穎的酒類就是Craft Mead。這種酒展示了一種挑戰性和自由性，表現出多樣化的口感體驗。",
-    "images": ["https://storage.googleapis.com/winentaste-alcohol-prod-assets/3154b03de77d4732a0ebf9dda484a8cb0cbeac6a.png?1704677992"]
-  },
-  {
-    "id": "基峰鶴",
-    "name": "基峰鶴",
-    "title": "",
-    "description": "基山商店位於九州佐賀縣的東端，靠近福岡縣的基山町，是一個小型的酒蔵。它的起源可以追溯到明治時期，當時附近的一些地主合作開始釀造清酒。基山商店利用來自九千部山和基山的伏流水以及基山豐富的自然資源，一直致力於為當地人提供美味的清酒​。基峰鶴是基山商店的主要品牌，名字的靈感來自基山的特別史跡基肄城，以及在基山山區自在飛舞的優雅鶴的形象​。在釀造清酒的過程中，基山商店特別注重原料的品質。他們100%使用被譽為酒米之最的山田錦，並且自30年前就開始與當地合約農家合作進行山田錦的栽培。從創業之初至今，基山商店始終保持著品質至上的態度，在釀酒過程中不斷追求卓越​​。",
-    "images": [""]
-  },
-  {
-    "id": "忠愛",
-    "name": "忠愛",
-    "title": "",
-    "description": "富川酒造是一家擁有超過100年歷史的日本清酒釀造廠，創立於大正2年（1913年）。最初，它是從事碾米業的，碾米是指將稻穀加工成精米的工藝，這也是釀造清酒的必要步驟。後來，富川酒造購入了酒造設施，並開始參與清酒的釀造業。目前，富川酒造的主要產品包括「忠愛」和「富美川」兩個品牌的清酒​。特別值得一提的是，富川酒造的「忠愛」系列清酒是使用日本百大名水之一的「尚仁沢」水源釀造的，這使得清酒味道特別純淨和鮮甜。而且，「忠愛」這個名字的由來很有意思，它是以「忠君愛國」的意念命名的，並且酒標上印有「忠愛」二字​。",
-    "images": [""]
-  },
-  {
-    "id": "白嶺",
-    "name": "白嶺",
-    "title": "ハクレイ酒造：有海的京都的釀酒藝術。",
-    "description": "自天保3年（1832年）創業以來，ハクレイ酒造一直為世界帶來頂級的日本酒。該酒蔵坐落於京都府宮津的丹後由良，這裡被譽為有海的京都。由良川靜靜地流經這裡，人們可以在河口近處欣賞到日本海的壯麗景色，而在其背後，是大江山連峰的由良ヶ岳，形成了一幅美麗的天然畫卷。在這樣的自然環境中，ハクレイ酒造從創業之初就堅持使用真名井的清澈泉水來釀製酒。而現在，他們帶來了新的純米吟醸無過濾生原酒系列，這不僅是對品質的追求，更是對釀酒傳統的創新和尊重。其中，品牌「Vivid」就是基於這樣的釀造哲學誕生的。當你首次看到它，最先引起注意的必定是瓶身上的標籤。這強烈而藝術性的視覺設計，是由跨足流行和藝術插畫的著名設計師安藤公三氏親手繪製。他的作品完美地融合了ハクレイ酒造的傳統與現代感，使每瓶酒都成為一件藝術品。",
-    "images": [""]
-  },
-  {
-    "id": "桂泉",
-    "name": "桂泉",
-    "title": "宮城之寶：くりこま高原的歷史釀酒藝術",
-    "description": "宮城縣的くりこま高原，這片土地不僅是稻米的故鄉，還擁有蒼翠的大自然景色，以及源源不絕的湧水恩賜。在這裡，我們所屬的酒造已有超過274年的歷史，追溯到1757年的江戸時代中期。我們的酒釀造歷史非常悠久。這裡豐富的稻米資源，湧泉的水質也是關鍵。事實上，明治天皇巡幸時，我們曾有幸獻上此地的湧泉水作為御前水。1978年，宮城県沖地震讓我們的酒蔵面臨倒塌的威脅，但這並沒有阻止我們前進。我們曾在南部杜氏的指導下，保持傳統的冷釀造酒。而宮城的女詩人原阿佐緒和演員菅原文太也曾與我們有深厚的緣分，他們在我們的酒窖中尋找靈感，慶祝新酒的誕生。現在，位於栗原市的釀造地，由第12代當主、音樂家かの香織帶領。她不受任何形式的束縛，釀造出搭配美食的最佳日本酒。自東日本大震災後，我們致力於支持宮城和整個日本酒行業，進行各種宣傳活動。作為國內最小的酒廠，我們的目標是傳承日本的文化，並以自己的步調前進。最後，我們的酒，它有著澄淨而不張揚的口感，帶著和平與舒適的味道。當你品嚐時，會感受到清新與淨化。我們期待能夠持續製作這樣的高質量、充滿愛與故事的酒，並與各位分享。",
-    "images": [""]
-  },
-  {
-    "id": "瑞冠",
-    "name": "瑞冠",
-    "title": "",
-    "description": "山岡酒造實踐著“伴農繁醸”，即與農業共同發展的釀酒方式，將“地酒是文化”作為其座右銘。酒造使用當地農家培育的亀の尾、山田錦、雄町、新千本等多種酒米，釀造出旨味豐富、味道深邃的瑞冠酒。自1980年代起，酒造還開始在自家農田種植酒米，採用無農藥的合鴨農法，致力於安全、可靠的酒米生產。山岡酒造積極與當地農業合作，促進循環型地區經濟發展。釀酒使用的水源是來自蔵地4公里外、冬季常有積雪的標高500米山區的“有田清水”，這種中軟水含有適量的礦物質，有助於酵母的健康發酵。特別是在低溫下進行的硝酸還原菌和乳酸菌發酵過程，這是自江戶時代以來傳承下來的酒母製作技術，對於傳統的きもと酒母製作來說，這樣的水源是不可或缺的。山岡酒造成立於大約1890年，起源於購買並接手經營自江戶時代以來的井堀酒場，現在是第四代經營者。創業初期以“祇園正宗”作為品牌名稱，但後來發現這個名稱侵犯了京都伏見的一家酒蔵的商標，於是在昭和2年改名為“瑞冠”，這是原本屬於四國一家酒蔵的名稱，寓意著“好事廣泛傳播”。山岡酒造位於廣島縣北部的三次市，這裡海拔350米，季節間溫差大，利於酒米生產。每年10月到次年4月的冬季，這裡低溫適合釀酒。",
-    "images": [""]
-  }
-]
