+++ src/pages/sakes/[sakeId].tsx
@@ -1,6 +1,10 @@
+import Typography from '@/components/common/Typography/Typography';
+import { TextTag } from '@/components/common/Typography/types';
 import type { GetStaticProps, GetStaticPaths } from 'next';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
 import { useRouter } from 'next/router';
+import styles from '@/components/common/Typography/Typography.module.css';
+import classNames from 'classnames';
 
 export const getStaticPaths: GetStaticPaths = async() => {
   return {
@@ -25,8 +29,10 @@ export default function Sakes() {
   const { sakeId } = router.query;
 
   return (
-    <div>
-      <h1>Sake ID: {sakeId}</h1>
+    <div className='flex w-full flex-col'>
+      <Typography className={classNames(styles['headline-sm'], 'mb-2 flex md:mb-3')} textTag={TextTag.H4}>
+        {sakeId}
+      </Typography>
     </div>
   );
 }
